chore: move activity, call to view-model.
