--- app/src/main/java/com/pereyrarg11/myquotes/MainActivity.kt
@@ -1,17 +1,0 @@
-package com.pereyrarg11.myquotes
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.pereyrarg11.myquotes.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
-    }
-}
