+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:id="@+id/main_container"
     android:background="@color/original_orange"
     android:padding="16dp"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <TextView
         android:id="@+id/quote"
